pyproject.toml
[tool.poetry]
name = "inference_server"
version = "0.1.0"
description = ""
authors = ["Matthew Chung <matthewchung74@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pip = "^21.1.3"
fastapi = "^0.65.1"
pydantic = "^1.8.2"
uvicorn = "^0.14.0"
asyncpg = "^0.23.0"
PyJWT = "^2.1.0"
python-dotenv = "^0.17.1"
PyGithub = "^1.55"
SQLAlchemy = "^1.4.19"
typer = "^0.3.2"
httpx = "^0.18.1"
boto3 = "^1.17.97"
docker = "^5.0.0"
nbconvert = "^6.0.7"
ipython = "^7.24.1"
aiofiles = "^0.7.0"
aio-pika = "^6.8.0"
psycopg2-binary = "^2.9.1"
yhat_params = {git = "https://github.com/yhatpub/yhat_params.git", rev = "main"}
opencv-python-headless = "^4.5.2"
Pillow = "^8.2.0"
tqdm = "^4.61.1"
websockets = "^9.1"
nest-asyncio = "^1.5.1"

[tool.poetry.dev-dependencies]
pytest-asyncio = "^0.15.1"
async-asgi-testclient = "^1.4.6"
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
